fix: add windows uts in pipeline for cns folder (#4206)
* move tests in cns/restserver that work only on linux to os specific files

* move tests in cns/kubecontroller that work only on linux to os specific files

changed vnet block request test case to expect vnetBlockNodeIP instead of primary ip to pass
in windows case

TODO: vnet block/prefix on nic tests need to be created for windows-- ideally the cases will
be moved back to conversion_test.go and the want section will change based on linux or windows

* fix cns/fsnotify tests to work on linux and windows without elevated permissions

* fix cns/stateprovider test running on windows

filepath, not path, should we used

* run routes_test on windows only if pipeline tag

* fix service_test.go and run new tests in pipeline

dpapi can only be imported in windows files

* remove trailing space


* remove fileContent

---------
