+++ cns/fsnotify/fsnotify_test.go
@@ -2,6 +2,7 @@ package fsnotify
 
 import (
 	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -23,7 +24,7 @@ func TestAddFile(t *testing.T) {
 			args: args{
 				podInterfaceID: "123",
 				containerID:    "67890",
-				path:           "/bad/path",
+				path:           "bad/path",
 			},
 			wantErr: true,
 		},
@@ -32,16 +33,23 @@ func TestAddFile(t *testing.T) {
 			args: args{
 				podInterfaceID: "345",
 				containerID:    "12345",
-				path:           "/path/we/want",
+				path:           "path/we/want",
 			},
 			wantErr: false,
 		},
 	}
+
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			err := os.MkdirAll("/path/we/want", 0o777)
+			baseDir := t.TempDir()
+
+			dirToCreate := filepath.Join(baseDir, "path", "we", "want")
+			err := os.MkdirAll(dirToCreate, 0o777)
 			require.NoError(t, err)
-			if err := AddFile(tt.args.podInterfaceID, tt.args.containerID, tt.args.path); (err != nil) != tt.wantErr {
+
+			fullPath := filepath.Join(baseDir, tt.args.path)
+
+			if err := AddFile(tt.args.podInterfaceID, tt.args.containerID, fullPath); (err != nil) != tt.wantErr {
 				t.Errorf("WatcherAddFile() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})
@@ -62,7 +70,7 @@ func TestWatcherRemoveFile(t *testing.T) {
 			name: "remove file fail",
 			args: args{
 				containerID: "12345",
-				path:        "/bad/path",
+				path:        "bad/path",
 			},
 			wantErr: true,
 		},
@@ -70,16 +78,23 @@ func TestWatcherRemoveFile(t *testing.T) {
 			name: "no such directory, add fail",
 			args: args{
 				containerID: "67890",
-				path:        "/path/we/want",
+				path:        "path/we/want",
 			},
 			wantErr: false,
 		},
 	}
+
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			err := os.MkdirAll("/path/we/want/67890", 0o777)
+			baseDir := t.TempDir()
+
+			dirToCreate := filepath.Join(baseDir, "path", "we", "want", "67890")
+			err := os.MkdirAll(dirToCreate, 0o777)
 			require.NoError(t, err)
-			if err := removeFile(tt.args.containerID, tt.args.path); (err != nil) != tt.wantErr {
+
+			fullPath := filepath.Join(baseDir, tt.args.path)
+
+			if err := removeFile(tt.args.containerID, fullPath); (err != nil) != tt.wantErr {
 				t.Errorf("WatcherRemoveFile() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})
