+++ cns/restserver/util_test.go
@@ -1,12 +1,9 @@
 package restserver
 
 import (
-	"context"
 	"testing"
 
-	"github.com/Azure/azure-container-networking/cns"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestAreNCsPresent(t *testing.T) {
@@ -161,22 +158,4 @@ func TestDeleteNCs(t *testing.T) {
 			assert.Equal(t, tt.want4, ncs)
 		})
 	}
-}
-
-func TestGetPnpIDMapping(t *testing.T) {
-	svc := getTestService(cns.KubernetesCRD)
-	svc.state.PnpIDByMacAddress = map[string]string{
-		"macaddress1": "value1",
-	}
-	pnpID, _ := svc.getPNPIDFromMacAddress(context.Background(), "macaddress1")
-	require.NotEmpty(t, pnpID)
-
-	// Backend network adapter not found
-	_, err := svc.getPNPIDFromMacAddress(context.Background(), "macaddress8")
-	require.Error(t, err)
-
-	// Empty pnpidmacaddress mapping
-	svc.state.PnpIDByMacAddress = map[string]string{}
-	_, err = svc.getPNPIDFromMacAddress(context.Background(), "macaddress8")
-	require.Error(t, err)
 }
