+++ cns/service_linux_test.go
@@ -1,0 +1,6 @@
+package cns
+
+// postProcessPEMCert does not do anything on linux
+func postProcessPEMCert(pem []byte) ([]byte, error) {
+	return pem, nil
+}
